split rmq_mgmt into rmq_mgmt and rmq_setup scripts. rmq_setup is script used only from command line. rmq_mgmt is a utility script for rmq management functions
